Implement variant generation based on graph and tries
